Refactor/api visibility (#42)
* refactor: change directory hierarchy of config

* refactor: change config.py to private

* refactor: change directory structure
