+++ src/majsoulrpa/__init__.py
@@ -1,7 +1,7 @@
 # ruff: noqa: F401
 from logging import NullHandler, getLogger
 
+from ._rpa import RPA
 from ._version import __version__
-from .rpa import RPA
 
 getLogger(__name__).addHandler(NullHandler())
