+++ src/majsoulrpa/common/_common.py
