+++ src/majsoulrpa/config/__init__.py
@@ -1,0 +1,2 @@
+# ruff: noqa: F401
+from ._config import get_config
