+++ src/majsoulrpa/config/authentication_imap.json
