+++ src/majsoulrpa/config/authentication_s3.json
