+++ src/majsoulrpa/config/schema.json
